src/backend/requirements.txt
# Backend dependencies - install with: pip install -r backend/requirements.txt
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
sqlalchemy>=2.0.45
pydantic[email]>=2.12.5
python-jose[cryptography]>=3.5.0
bcrypt>=4.0.0
python-multipart>=0.0.21
python-dotenv>=1.2.1
psycopg2-binary>=2.9.9
